Base of the project
Created a basic GET request handler for /chat endpoint. This
currently has almost no real functionality, but it is a good
base. The entirely of this project's network code will be
done with ktor.

Made a simple Dockerfile

Just to get this project off the ground and into a container.